MoleculeProjectEntry: add "op" typed
